Update wordle_solution.txt
It turns out I understood Wordle's scoring algorithm wrong. If you guess 'AXAXA' and the secret word is 'BBXBB' then only the first 'X' is identified as matching but in the incorrect place. I've updated the scoring algorithm and it now makes the correct choices to win todays (#204) game of Wordle.